- simpified getAlphas() method. - Ribosome movement and updating of mrna elements is done manually by the gillespie algorithm.
This is a big and complex commit. I've been testing it for quite a while but seems to be doing
fine. Ribosome movement proved to be quite complex (number of corner cases). A alternative
approach is to change the enlongating and terminating objects to get a pointer to the next codon
and then get the alphas based on that pointer.